requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
onnxruntime==1.17.1
numpy==1.26.4
pillow==10.2.0